Added test files for HMM generation. generate_HMM is now running and sucessfully creating HMMs for a small toy file (unicorn_train.tsv) and a small exerpt (about 50400 lines) from the Wall Street Journal (small_wsj_train.tsv).
Next step: Add metadata section at the beginning of the generated file 
that includes the number of states and symbols. This will make it easier
to read in the HMM later on, enabling the internal data structures to 
allocate the required capacity all at once, and not step by step.